leere gruppen können nicht abgespeichert werden - fehler weiterhin vorhanden

+++ src/dao/GroupDaoImpl.java
@@ -33,17 +33,28 @@ public class GroupDaoImpl implements GroupDao  {
 			pstmt = connection.prepareStatement("delete from usertogroup where groupid = ?");
 			pstmt.setLong(1, group);
 			pstmt.executeUpdate();
-			
-			for (String user : userList) {
-				Long userID = Long.parseLong(user);
-				pstmt = connection.prepareStatement("delete from usertogroup where userid = ?");
-				pstmt.setLong(1, userID);
-				pstmt.executeUpdate();
+			if (group == 1) {
 				pstmt = connection.prepareStatement("insert into usertogroup (userid, groupid) values (?, ?)");
-				pstmt.setLong(1, userID);
-				pstmt.setLong(2, group);
+				pstmt.setLong(1, 1);
+				pstmt.setLong(2, 1);
 				pstmt.executeUpdate();
 			}
+			
+			System.out.println(userList.length);
+
+			if(userList.length != 0 ){
+				for (String user : userList) {
+					Long userID = Long.parseLong(user);
+					pstmt = connection.prepareStatement("delete from usertogroup where userid = ?");
+					pstmt.setLong(1, userID);
+					pstmt.executeUpdate();
+					pstmt = connection.prepareStatement("insert into usertogroup (userid, groupid) values (?, ?)");
+					pstmt.setLong(1, userID);
+					pstmt.setLong(2, group);
+					pstmt.executeUpdate();
+				}
+			}
+
 		} catch (Exception e) {
 			throw new GroupNotSavedException();
 		} finally {
@@ -109,7 +120,9 @@ public class GroupDaoImpl implements GroupDao  {
 					userInGroup.setName(rsAllUser.getString("username"));
 					userInGroup.setAccess(access);
 					userInGroup.setUserid(rsAllUser.getLong("id"));
-					groupList.add(userInGroup);
+					if(userInGroup.getUserid() != 1){
+						groupList.add(userInGroup);
+					}
 					access = 0;
 				}				
 			return groupList;
